+++ DarkLand/Assets/Scripts/WaypointManager/WaypointMover.cs
@@ -15,7 +15,7 @@ public class WaypointMover : MonoBehaviour
     private bool alive = false;
 
     private bool isAttacking = false;
-    [SerializeField] private float attackDamage = 20;
+    [SerializeField] private int attackDamage = 20;
     private Transform currentWaypoint = null;
     // Start is called before the first frame update
     void Start(){
